+++ app/src/main/java/com/youneshatti/jarboa/data/local/DatabaseConverters.kt
@@ -1,6 +1,7 @@
 package com.youneshatti.jarboa.data.local
 
 import androidx.room.TypeConverter
+import com.youneshatti.jarboa.domain.model.MessageEncryption
 import com.youneshatti.jarboa.domain.model.MessageStatus
 
 class DatabaseConverters {
@@ -9,5 +10,10 @@ class DatabaseConverters {
 
     @TypeConverter
     fun stringToMessageStatus(value: String): MessageStatus = MessageStatus.valueOf(value)
-}
 
+    @TypeConverter
+    fun messageEncryptionToString(encryption: MessageEncryption): String = encryption.name
+
+    @TypeConverter
+    fun stringToMessageEncryption(value: String): MessageEncryption = MessageEncryption.valueOf(value)
+}